[feat] 여행 동행 선택화면 구현
