+++ feature/trip-list/src/main/kotlin/com/tripmate/android/feature/trip_list/viewmodel/TripListUiAction.kt
@@ -3,4 +3,5 @@ package com.tripmate.android.feature.trip_list.viewmodel
 sealed interface TripListUiAction {
     data object OnBackClicked : TripListUiAction
     data class OnTabChanged(val index: Int) : TripListUiAction
+    data class OnTicketClicked(val ticketId: Int) : TripListUiAction
 }
